Create new hook for fetching registration
